tests/requirements.txt
pytest==6.2.0
pytest-aiohttp==0.3.0
